+++ R/make_read.R
@@ -19,12 +19,13 @@
 #' @param sv Allows users to specify whether their csv file is
 #' `"comma"` (English) or `"semi-colon"` (European) separated.
 #' @param ... Additional parameters passed to the read/write function.
-#' @return The `read_edgelist()` and `read_nodelist()` functions will import
+#' @return `read_edgelist()` and `read_nodelist()` will import
 #' into edgelist (tibble) format which can then be coerced or combined into
 #' different graph objects from there.
 #'
-#' The `read_pajek()` and `read_ucinet()` functions will import into
+#' `read_pajek()` and `read_ucinet()` will import into
 #' a tidygraph format, since they already contain both edge and attribute data.
+#' `read_matrix()` will import into tidygraph format too.
 #' Note that all graphs can be easily coerced into other formats
 #' with `{migraph}`'s `as_` methods.
 #'
@@ -65,6 +66,34 @@
 #' @name read
 #' @seealso [as]
 NULL
+
+#' @describeIn read Reading adjacency matrices from Excel/csv files
+#' @export
+read_matrix <- function(file = file.choose(),
+                          sv = c("comma", "semi-colon"),
+                          ...) {
+  sv <- match.arg(sv)
+  if (grepl("csv$", file)) {
+    if (sv == "comma") {
+      out <- read.csv(file, ...) # For US
+    } else {
+      out <- read.csv2(file, ...) # For EU
+    }
+  } else if (grepl("xlsx$|xls$", file)) {
+    if(requireNamespace("readxl", quietly = TRUE)){
+      out <- readxl::read_excel(file, ...)  
+    } else stop("Please install `readxl` from CRAN to import Excel files.")
+  }
+  if((dim(out)[1]+1) == dim(out)[2])
+    out <- out[,-1]
+  if(!is.null(colnames(out)) & 
+     all(colnames(out) == paste0("X",seq_along(colnames(out)))))
+    colnames(out) <- NULL
+  if(!is.null(colnames(out)) & is.null(rownames(out)) &
+     dim(out)[1] == dim(out)[2])
+    rownames(out) <- colnames(out)
+  as_tidygraph(as.matrix(out))
+}
 
 #' @describeIn read Reading edgelists from Excel/csv files
 #' @export
@@ -499,3 +528,63 @@ write_ucinet <- function(object,
   }
   close(UCINET.data)
 }
+
+#' @describeIn read Reading DynetML files
+#' @export
+read_dynetml <- function(file = file.choose()) {
+  if(!requireNamespace("xml2", quietly = TRUE)){
+    stop("Please install `xml2` from CRAN to import DynetML files.")
+  } else {
+    
+    name <- type <- nodeset <- target <- value <- NULL
+    
+    xmlfile <- xml2::read_xml(file)
+    xmllist <- xml2::as_list(xmlfile)
+    
+    # Getting nodeset
+    # to deal with legacy constructions:
+    if("MetaMatrix" %in% names(xmllist$DynamicNetwork))
+      nodesets <- xmllist$DynamicNetwork$MetaMatrix$nodes else
+        nodesets <- xmllist$DynamicNetwork$MetaNetwork$nodes
+    nodesets <- dplyr::coalesce(unlist(lapply(nodesets, 
+                                              function(x) ifelse(is.null(attr(x, "id")),
+                                                                 NA_character_, attr(x, "id")))),
+                                unlist(lapply(nodesets, 
+                                              function(x) ifelse(is.null(attr(x, "type")),
+                                                                 NA_character_, attr(x, "type")))))
+    # to deal with legacy constructions:
+    if("MetaMatrix" %in% names(xmllist$DynamicNetwork)){
+      nodesets <- unname(rep(nodesets, vapply(xmllist$DynamicNetwork$MetaMatrix$nodes,
+                                              function(x) length(x), numeric(1))))
+    } else
+      nodesets <- unname(rep(nodesets, vapply(xmllist$DynamicNetwork$MetaNetwork$nodes,
+                                              function(x) length(x), numeric(1)))) 
+    
+    # Getting nodes
+    nodes <- xml2::as_list(xml2::xml_find_all(xmlfile, ".//node"))
+    nodes <- dplyr::bind_rows(lapply(nodes, function(x){
+      values <- sapply(x$properties, function(y) attr(y, "value"))
+      attrs <- sapply(x$properties, function(y) attr(y, "name"))
+      names(values) <- attrs
+      c(name = attr(x, "id"), values)
+    }))
+    # Add nodeset information if necessary
+    if(length(unique(nodesets))==2)
+      nodes <- nodes %>% dplyr::mutate(type = nodesets == unique(nodesets)[2]) %>% 
+      dplyr::select(name, type, dplyr::everything()) else if (length(unique(nodesets))>2)
+        nodes <- nodes %>% dplyr::mutate(nodeset = nodesets) %>% 
+      dplyr::select(name, nodeset, dplyr::everything())
+    
+    # Getting edges
+    edgelist <- xml2::xml_attrs(xml2::xml_find_all(xmlfile, ".//edge"))
+    # to deal with legacy constructions:
+    if(length(edgelist)==0) edgelist <- xml2::xml_attrs(xml2::xml_find_all(xmlfile, ".//link"))
+    edgelist <- as.data.frame(t(sapply(edgelist, function(x) x, simplify = TRUE)))
+    edgelist$type <- NULL
+    edgelist$value <- as.numeric(edgelist$value)
+    edgelist <- dplyr::filter(edgelist, source %in% nodes$name & target %in% nodes$name)
+    edgelist <- dplyr::filter(edgelist, value != 0)
+    as_tidygraph(list(nodes = nodes, ties = edgelist))
+  }
+}
+